Codefactor:
- Split out new gradient tests to avoid codefactor complaining about the module being too long
- Break down einsumLinalgGeneric to reduce single-function complexiy
